character stat rework

+++ Main Project/Assets/Scripts/StatSystem/StatBonus.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class StatBonus
 {
-    public int BonusValue { get; set; }
+    public float BonusValue { get; set; }
 
-    public StatBonus(int bonusValue)
+    public StatBonus(float bonusValue)
     {
         this.BonusValue = bonusValue;
     }
